refactor: refactored customer, wallet and payout requests and responses

+++ Embedly.SDK/Models/Responses/Wallets/Wallet.cs
@@ -30,13 +30,13 @@ public sealed class Wallet
     ///     Gets or sets the available balance.
     /// </summary>
     [JsonPropertyName("availableBalance")]
-    public double AvailableBalance { get; set; }
+    public decimal AvailableBalance { get; set; }
 
     /// <summary>
     ///     Gets or sets the ledger balance.
     /// </summary>
     [JsonPropertyName("ledgerBalance")]
-    public double LedgerBalance { get; set; }
+    public decimal LedgerBalance { get; set; }
 
     /// <summary>
     ///     Gets or sets the wallet restriction ID.
@@ -78,7 +78,7 @@ public sealed class Wallet
     ///     Gets or sets the overdraft limit.
     /// </summary>
     [JsonPropertyName("overdraft")]
-    public double? Overdraft { get; set; }
+    public decimal? Overdraft { get; set; }
 
     /// <summary>
     ///     Gets or sets the virtual account information.
